Change PanelController into abstract class
